requirements.txt
adrf==0.1.9
aioredis==2.0.1
app-store-server-library==1.9.0
asyncio==3.4.3
boto3==1.39.12
channels==4.2.2
channels-redis==4.3.0
cryptography==45.0.5
dj-database-url==2.2.0
Django==5.2.4
django-stubs==5.2.2
django-stubs-ext==5.2.2
django-allauth==65.10.0
django-cors-headers==4.7.0
django-oauth-toolkit==2.4.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
hiredis==3.2.1
httpx[http2]==0.28.1
openai==1.97.1
openai-agents==0.2.3
pgvector==0.4.1
pretty_errors==1.2.25
psycopg[binary]==3.2.9
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
redis==5.3.0
sendgrid==6.12.4
sentry-sdk[django]==2.33.2
stripe==12.3.0
taskiq[reload]==0.11.18
taskiq-redis==1.0.9
twilio==9.7.0
uvicorn[standard]==0.35.0
markdown==3.8.2
langfuse==3.2.1
